+++ examples/hello-world.ts
@@ -1,1 +1,1 @@
-console.log(`Hello, World!`);
+console.log("Hello, World!");
