+++ src/argparse/parser.ts
@@ -2,7 +2,8 @@ import fs from "node:fs";
 import { resolveFilesAndDirectoriesToFiles } from "@/argparse/utils.js";
 import { analyze } from "@/commands/analyze.js";
 import { ExitCodesEnum } from "@/constants.js";
-import { LogLevelsEnum, logger, setLogLevel } from "@/logger.js";
+import { LogLevelsEnum } from "@/loggers/common.js";
+import { logger, setLogLevel } from "@/loggers/program.js";
 import { SeverityLevelsEnum } from "@/model.js";
 import { hideBin } from "yargs/helpers";
 import yargs from "yargs/yargs";
