--- src/logger.ts
@@ -1,28 +1,0 @@
-import winston from "winston";
-
-export enum LogLevelsEnum {
-  ERROR = "error",
-  WARNING = "warning",
-  INFO = "info",
-  DEBUG = "debug",
-}
-
-export const logger = winston.createLogger({
-  levels: winston.config.syslog.levels,
-  format: winston.format.align(),
-  transports: [
-    new winston.transports.Console({
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.simple(),
-      ),
-      level: "debug",
-    }),
-  ],
-});
-
-export function setLogLevel(level: LogLevelsEnum) {
-  for (const transport of logger.transports) {
-    transport.level = level;
-  }
-}
