+++ src/loggers/common.ts
@@ -1,0 +1,6 @@
+export enum LogLevelsEnum {
+  ERROR = "error",
+  WARNING = "warning",
+  INFO = "info",
+  DEBUG = "debug",
+}
