+++ tsconfig.json
@@ -7,6 +7,8 @@
   ],
   "compilerOptions": {
     "outDir": "dist",
+    "sourceMap": true,
+    "inlineSourceMap": true,
     "resolveJsonModule": true,
     "incremental": true, // allow `tsc` to persist cache to `.tsbuildinfo` for faster rebuilds
     "baseUrl": ".",
